requirements.txt
torch==2.4.0
torchvision==0.15.0
opencv-python==4.5.5.64
numpy>=1.20.0
matplotlib>=3.5.0
tqdm>=4.62.0
pillow>=9.0.0
scikit-learn>=1.0.0
pandas>=1.3.0
tensorboard>=2.8.0
scipy>=1.7.0
seaborn>=0.11.0
ipywidgets>=7.6.0
jupyter>=1.0.0
ipykernel>=6.0.0
notebook>=6.4.0
pytest>=7.0.0
mypy>=0.910
flake8>=4.0.0
black>=22.0.0
isort>=5.10.0
flask>=2.0.0
werkzeug>=2.0.0
jinja2>=3.0.0
itsdangerous>=2.0.0
click>=8.0.0 